feature/28: refactoring stock
