+++ product/src/main/java/takeoff/logistics_service/msa/product/stock/application/service/StockServiceImpl.java
@@ -19,7 +19,7 @@ import takeoff.logistics_service.msa.product.stock.presentation.dto.request.Incr
 import takeoff.logistics_service.msa.product.stock.presentation.dto.request.PostStockRequestDto;
 import takeoff.logistics_service.msa.product.stock.presentation.dto.request.PrepareStockRequestDto;
 import takeoff.logistics_service.msa.product.stock.presentation.dto.request.StockItemRequestDto;
-import takeoff.logistics_service.msa.product.stock.presentation.dto.request.StockSearchCondition;
+import takeoff.logistics_service.msa.product.stock.application.dto.StockSearchCondition;
 import takeoff.logistics_service.msa.product.stock.presentation.dto.response.DecreaseStockResponseDto;
 import takeoff.logistics_service.msa.product.stock.presentation.dto.response.GetStockResponseDto;
 import takeoff.logistics_service.msa.product.stock.presentation.dto.response.IncreaseStockResponseDto;
