+++ product/src/main/java/takeoff/logistics_service/msa/product/stock/infrastructure/persistence/impl/JpaStockRepositoryCustomImpl.java
@@ -14,7 +14,7 @@ import org.springframework.data.domain.PageImpl;
 import org.springframework.data.domain.Pageable;
 import takeoff.logistics_service.msa.product.stock.infrastructure.persistence.JpaStockRepositoryCustom;
 import takeoff.logistics_service.msa.product.stock.presentation.dto.StockIdDto;
-import takeoff.logistics_service.msa.product.stock.presentation.dto.request.StockSearchCondition;
+import takeoff.logistics_service.msa.product.stock.application.dto.StockSearchCondition;
 import takeoff.logistics_service.msa.product.stock.presentation.dto.response.GetStockResponseDto;
 
 @RequiredArgsConstructor
