+++ product/src/main/java/takeoff/logistics_service/msa/product/stock/model/repository/StockRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import takeoff.logistics_service.msa.product.stock.model.entity.Stock;
 import takeoff.logistics_service.msa.product.stock.model.entity.StockId;
-import takeoff.logistics_service.msa.product.stock.presentation.dto.request.StockSearchCondition;
+import takeoff.logistics_service.msa.product.stock.application.dto.StockSearchCondition;
 import takeoff.logistics_service.msa.product.stock.presentation.dto.response.GetStockResponseDto;
 
 public interface StockRepository {
